+++ ApptentiveConnect/source/Engagement/Model/ATEngagementManifestParser.m
@@ -12,21 +12,19 @@
 #import "ATInteraction.h"
 #import <UIKit/UIKit.h>
 #import "ATConnect_Debugging.h"
+#import "ATInteractionInvocation.h"
 
 @implementation ATEngagementManifestParser
 
-- (NSDictionary *)codePointInteractionsForEngagementManifest:(NSData *)jsonManifest {
+
+
+- (NSDictionary *)targetsAndInteractionsForEngagementManifest:(NSData *)jsonManifest {
+	// JSON String for testing. Using "Copy" command on variable in debugger preserves escape characters.
+	//NSString *jsonString = [[[NSString alloc] initWithData:jsonManifest encoding:NSUTF8StringEncoding] autorelease];
 	
-	if ([ATConnect sharedConnection].debuggingOptions & ATConnectDebuggingOptionsLogAllHTTPRequests) {
-		NSString* interactionsString = [[NSString alloc] initWithData:jsonManifest encoding:NSUTF8StringEncoding];
-		
-		ATLogDebug(@"Retrieved Apptentive Interaction JSON from server:");
-		ATLogDebug(@"%@", interactionsString);
-		
-		[interactionsString release];
-	}
+	NSMutableDictionary *targets = [NSMutableDictionary dictionary];
+	NSMutableDictionary *interactions = [NSMutableDictionary dictionary];
 	
-	NSDictionary *codePointInteractions = nil;
 	BOOL success = NO;
 	
 	@autoreleasepool {
@@ -35,23 +33,24 @@
 			
 			id decodedObject = [ATJSONSerialization JSONObjectWithData:jsonManifest error:&error];
 			if (decodedObject && [decodedObject isKindOfClass:[NSDictionary class]]) {
-				success = YES;
-				NSDictionary *jsonManifest = (NSDictionary *)decodedObject;
-				NSDictionary *jsonCodePoints = [jsonManifest objectForKey:@"interactions"];
+				NSDictionary *jsonDictionary = (NSDictionary *)decodedObject;
 				
-				NSMutableDictionary *codePoints = [[NSMutableDictionary alloc] init];
-				for (NSString *codePointName in [jsonCodePoints allKeys]) {
-					NSArray *jsonInteractions = [jsonCodePoints objectForKey:codePointName];
-					
-					NSMutableArray *interactions = [NSMutableArray array];
-					for (NSDictionary *jsonInteraction in jsonInteractions) {
-						ATInteraction *interaction = [ATInteraction interactionWithJSONDictionary:jsonInteraction];
-						[interactions addObject:interaction];
-					}
-					[codePoints setObject:interactions forKey:codePointName];
+				// Targets
+				NSDictionary *targetsDictionary = jsonDictionary[@"targets"];
+				for (NSString *event in [targetsDictionary allKeys]) {
+					NSArray *invocationsJSONArray = targetsDictionary[event];
+					NSArray *invocationsArray = [ATInteractionInvocation invocationsWithJSONArray:invocationsJSONArray];
+					[targets setObject:invocationsArray forKey:event];
 				}
 				
-				codePointInteractions = codePoints;
+				// Interactions
+				NSArray *interactionsArray = jsonDictionary[@"interactions"];
+				for (NSDictionary *interactionDictionary in interactionsArray) {
+					ATInteraction *interactionObject = [ATInteraction interactionWithJSONDictionary:interactionDictionary];
+					[interactions setObject:interactionObject forKey:interactionObject.identifier];
+				}
+				
+				success = YES;
 			} else {
 				[parserError release], parserError = nil;
 				parserError = [error retain];
@@ -63,12 +62,14 @@
 			success = NO;
 		}
 	}
-	if (!success) {
-		codePointInteractions = nil;
-	} else {
-		[codePointInteractions autorelease];
+	
+	NSDictionary *targetsAndInteractions = nil;
+	if (success) {
+		targetsAndInteractions = @{@"targets": targets,
+								   @"interactions": interactions};
 	}
-	return codePointInteractions;
+	
+	return targetsAndInteractions;
 }
 
 - (NSError *)parserError {