+++ ApptentiveConnect/source/Engagement/Persistence/ATEngagementBackend.h
@@ -23,43 +23,44 @@ NSString *const ATEngagementInteractionsInvokesVersionKey;
 NSString *const ATEngagementInteractionsInvokesBuildKey;
 NSString *const ATEngagementInteractionsInvokesLastDateKey;
 
+NSString *const ATEngagementCodePointHostAppVendorKey;
+NSString *const ATEngagementCodePointHostAppInteractionKey;
+NSString *const ATEngagementCodePointApptentiveVendorKey;
+NSString *const ATEngagementCodePointApptentiveAppInteractionKey;
+
 @class ATInteraction;
 
 @interface ATEngagementBackend : NSObject {
 @private
-	NSMutableDictionary *codePointInteractions;
+	NSMutableDictionary *_engagementTargets;
+	NSMutableDictionary *_engagementInteractions;
 }
 
 + (ATEngagementBackend *)sharedBackend;
 
 - (void)checkForEngagementManifest;
 - (BOOL)shouldRetrieveNewEngagementManifest;
-- (void)didReceiveNewCodePointInteractions:(NSDictionary *)codePointInteractions maxAge:(NSTimeInterval)expiresMaxAge;
-- (void)updateVersionInfo;
-+ (NSString *)cachedEngagementStoragePath;
 
-- (NSArray *)interactionsForCodePoint:(NSString *)codePoint;
-- (ATInteraction *)interactionForCodePoint:(NSString *)codePoint;
+- (void)didReceiveNewTargets:(NSDictionary *)targets andInteractions:(NSDictionary *)interactions maxAge:(NSTimeInterval)expiresMaxAge;
+
+- (void)updateVersionInfo;
++ (NSString *)cachedTargetsStoragePath;
++ (NSString *)cachedInteractionsStoragePath;
+
+- (ATInteraction *)interactionForEvent:(NSString *)event;
+
+- (ATInteraction *)interactionForInvocations:(NSArray *)invocations;
 
 - (BOOL)willShowInteractionForLocalEvent:(NSString *)event;
 - (BOOL)willShowInteractionForCodePoint:(NSString *)codePoint;
 
 + (NSString *)stringByEscapingCodePointSeparatorCharactersInString:(NSString *)string;
-+ (NSString *)codePointForLocalEvent:(NSString *)event;
-+ (NSString *)codePointForVendor:(NSString *)vendor interaction:(NSString *)interaction event:(NSString *)event;
-	
-- (BOOL)engageLocalEvent:(NSString *)eventLabel fromViewController:(UIViewController *)viewController;
-- (BOOL)engageLocalEvent:(NSString *)eventLabel userInfo:(NSDictionary *)userInfo customData:(NSDictionary *)customData extendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
++ (NSString *)codePointForVendor:(NSString *)vendor interactionType:(NSString *)interactionType event:(NSString *)event;
 
-- (BOOL)engageApptentiveAppEvent:(NSString *)eventLabel userInfo:(NSDictionary *)userInfo;
-- (BOOL)engageApptentiveEvent:(NSString *)eventLabel fromInteraction:(ATInteraction *)interaction fromViewController:(UIViewController *)viewController;
-- (BOOL)engageApptentiveEvent:(NSString *)eventLabel fromInteraction:(ATInteraction *)interaction fromViewController:(UIViewController *)viewController userInfo:(NSDictionary *)userInfo;
+- (BOOL)engageApptentiveAppEvent:(NSString *)event;
+- (BOOL)engageLocalEvent:(NSString *)event userInfo:(NSDictionary *)userInfo customData:(NSDictionary *)customData extendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
 
-- (BOOL)engageEvent:(NSString *)eventLabel fromVendor:(NSString *)vendor fromInteraction:(NSString *)interaction userInfo:(NSDictionary *)userInfo fromViewController:(UIViewController *)viewController;
-- (BOOL)engageEvent:(NSString *)eventLabel fromVendor:(NSString *)vendor fromInteraction:(NSString *)interaction userInfo:(NSDictionary *)userInfo customData:(NSDictionary *)customData extendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
-
-- (BOOL)engage:(NSString *)codePoint userInfo:(NSDictionary *)userInfo fromViewController:(UIViewController *)viewController;
-- (BOOL)engage:(NSString *)codePoint userInfo:(NSDictionary *)userInfo customData:(NSDictionary *)customData extendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
+- (BOOL)engageCodePoint:(NSString *)codePoint fromInteraction:(ATInteraction *)fromInteraction userInfo:(NSDictionary *)userInfo customData:(NSDictionary *)customData extendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
 
 - (void)codePointWasSeen:(NSString *)codePoint;
 - (void)codePointWasEngaged:(NSString *)codePoint;
@@ -69,6 +70,8 @@ NSString *const ATEngagementInteractionsInvokesLastDateKey;
 - (void)presentInteraction:(ATInteraction *)interaction fromViewController:(UIViewController *)viewController;
 - (void)presentUpgradeMessageInteraction:(ATInteraction *)interaction fromViewController:(UIViewController *)viewController;
 - (void)presentEnjoymentDialogInteraction:(ATInteraction *)interaction fromViewController:(UIViewController *)viewController;
+
+- (void)presentNavigateToLinkInteraction:(ATInteraction *)interaction;
 
 // Used for debugging only.
 - (void)resetUpgradeVersionInfo;